some polishing of headers signing - still WIP
