+++ core.js
@@ -261,82 +261,4 @@ function hash(string, encoding) {
     return crypto.sha256(string, encoding);
 }
 
-function fillOptions(options) {
-    options = options || {};
-    options.key = options.key || __ENV.AWS_ACCESS_KEY_ID;
-    options.secret = options.secret || __ENV.AWS_SECRET_ACCESS_KEY;
-    options.sessionToken = options.sessionToken || __ENV.AWS_SESSION_TOKEN;
-    options.protocol = options.protocol || "https";
-    options.timestamp = options.timestamp || Date.now();
-    options.region = options.region || __ENV.AWS_REGION || "us-east-1";
-    options.expires = options.expires || 86400; // 24 hours
-    options.headers = options.headers || {};
-    options.signSessionToken = options.signSessionToken || false;
-    options.doubleEscape =
-        options.doubleEscape !== undefined ? options.doubleEscape : true;
-    return options;
-}
 
-function request(method, service, region, target, path="", body=null, query="", params= {}) {
-    var options =  {};
-    options.key = options.key || __ENV.AWS_ACCESS_KEY_ID;
-    options.secret = options.secret || __ENV.AWS_SECRET_ACCESS_KEY;
-    options.protocol = "https";
-    options.method = method;
-    options.signSessionToken = true;
-    options.doubleEscape = false;
-    options.headers = {}
-    options.timestamp = options.timestamp || Date.now();
-    if (typeof params !== "undefined" && typeof params.headers !== "undefined") {
-        options.headers = params.headers;
-    }
-    options.headers["X-Amz-Target"] = target;
-    options.headers["Content-Type"] = "application/x-amz-json-1.1";
-    var credential = options.key +
-        "/" +
-        createCredentialScope(options.timestamp, region, service);
-    var signedHeaders = createSignedHeaders(options.headers);
-    var signature = generateSignature(service, method, path, query, body, options);
-    var signatureHeader = "AWS4-HMAC-SHA256 Credential=".concat(
-        credential, 
-        " SignedHeaders=" ,
-        signedHeaders,
-        " Signature=",
-        signature,
-    )
-
-    options.headers["Authorization"] = signatureHeader;
-    params.headers = options.headers
-    var url = options.protocol.concat( "://", service ,".", region, ".amazonaws.com")
-
-    return http.request(method, url, body , params)
-}
-exports.request=request;
-
-function generateSignature(service, method, path ,query, body, options) {
-
-    var canonicalRequest = createCanonicalRequest(
-        method,
-        path,
-        query,
-        options.headers,
-        body,
-        options.doubleEscape
-    );
-
-    var stringToSign = createStringToSign(
-        options.timestamp,
-        options.region,
-        service,
-        canonicalRequest
-    );
-
-    var signature = createSignature(
-        options.secret,
-        options.timestamp,
-        options.region,
-        service,
-        stringToSign
-    );
-    return signature
-}
