+++ docker-compose.yaml
@@ -1,0 +1,13 @@
+version: '2.1'
+
+services:
+  localstack:
+    container_name: "${LOCALSTACK_DOCKER_NAME-localstack_main}"
+    image: localstack/localstack:0.12.2
+    ports:
+      - "4566-4599:4566-4599"
+      - "${PORT_WEB_UI-8080}:${PORT_WEB_UI-8080}"
+    environment:
+      - SERVICES=s3,secretsmanager
+      - DEBUG=1
+      - SECRETSMANAGER_PORT=4599
