+++ s3.js
@@ -1,4 +1,4 @@
-var createdPresignedURL = require(./core.js).createPresignedURL;
+var createPresignedURL = require("./core.js").createPresignedURL;
 
 function createPresignedS3URL(name, options) {
   options = options || {};
