+++ secretsManager.js
@@ -1,3 +1,4 @@
+var http = require("k6/http");
 var aws = require("./core.js");
 
 function getSecret(region, keyname, params = {}) {
